requirements.txt
fastapi~=0.97.0
pydantic~=1.10.9
motor~=3.2.0
python-dotenv~=1.0.0
pymongo~=4.4.0
tenacity~=8.2.2